+++ examples/blog/hackathonize/src/App.tsx
@@ -82,7 +82,7 @@ function App() {
             notificationProvider={notificationProvider}
             Layout={Layout}
             catchAll={<ErrorComponent />}
-            disableTelemetry={true}
+            options={{ disableTelemetry: true }}
         />
     );
 }