+++ examples/blog/refeedback/src/App.tsx
@@ -39,7 +39,7 @@ function App() {
             notificationProvider={notificationProvider}
             LoginPage={LoginPage}
             catchAll={<ErrorComponent />}
-            disableTelemetry={true}
+            options={{ disableTelemetry: true }}
         />
     );
 }