+++ examples/i18n/nextjs/pages/_app.tsx
@@ -6,7 +6,6 @@ import { appWithTranslation, useTranslation } from "next-i18next";
 
 import {
     notificationProvider,
-    LoginPage,
     Layout,
     ErrorComponent,
 } from "@pankod/refine-antd";
@@ -47,7 +46,7 @@ function MyApp({ Component, pageProps }: AppProps): JSX.Element {
             notificationProvider={notificationProvider}
             Layout={Layout}
             catchAll={<ErrorComponent />}
-            disableTelemetry={true}
+            options={{ disableTelemetry: true }}
         >
             <Component {...pageProps} />
         </Refine>