+++ examples/table/mui/advancedTable/src/App.tsx
@@ -50,7 +50,7 @@ const App: React.FC = () => {
                             options: { route: "data-grid", label: "Data Grid" },
                         },
                     ]}
-                    disableTelemetry={true}
+                    options={{ disableTelemetry: true }}
                 />
             </RefineSnackbarProvider>
         </ThemeProvider>