+++ examples/web3/ethereumLogin/src/App.tsx
@@ -35,7 +35,7 @@ function App() {
             notificationProvider={notificationProvider}
             Layout={Layout}
             catchAll={<ErrorComponent />}
-            disableTelemetry={true}
+            options={{ disableTelemetry: true }}
         />
     );
 }