+++ packages/core/src/contexts/refine/IRefineContext.ts
@@ -1,12 +1,46 @@
 import { RefineProps } from "@components/containers";
 import React, { ReactNode } from "react";
+import { QueryClientConfig } from "@tanstack/react-query";
+import { ReactQueryDevtools } from "@tanstack/react-query-devtools";
 
 import {
     MutationMode,
     TitleProps,
     LayoutProps,
     LiveModeProps,
+    RedirectionTypes,
 } from "../../interfaces";
+
+export interface IRefineOptions {
+    mutationMode?: MutationMode;
+    syncWithLocation?: boolean;
+    warnWhenUnsavedChanges?: boolean;
+    undoableTimeout?: number;
+    liveMode?: LiveModeProps["liveMode"];
+    disableTelemetry?: boolean;
+    redirect?: {
+        afterCreate?: RedirectionTypes;
+        afterClone?: RedirectionTypes;
+        afterEdit?: RedirectionTypes;
+    };
+    reactQuery?: {
+        clientConfig?: QueryClientConfig;
+        devtoolConfig?: React.ComponentProps<typeof ReactQueryDevtools> | false;
+    };
+}
+
+export interface IRefineContextOptions {
+    mutationMode: MutationMode;
+    syncWithLocation: boolean;
+    warnWhenUnsavedChanges: boolean;
+    undoableTimeout: number;
+    liveMode: LiveModeProps["liveMode"];
+    redirect: {
+        afterCreate: RedirectionTypes;
+        afterClone: RedirectionTypes;
+        afterEdit: RedirectionTypes;
+    };
+}
 
 export interface IRefineContext {
     hasDashboard: boolean;
@@ -23,8 +57,9 @@ export interface IRefineContext {
     Header?: React.FC;
     Footer?: React.FC;
     OffLayoutArea?: React.FC;
-    liveMode?: LiveModeProps["liveMode"];
+    liveMode: LiveModeProps["liveMode"];
     onLiveEvent?: LiveModeProps["onLiveEvent"];
+    options: IRefineContextOptions;
 }
 
 export interface IRefineContextProvider {
@@ -42,7 +77,8 @@ export interface IRefineContextProvider {
     Header?: React.FC;
     Footer?: React.FC;
     OffLayoutArea?: React.FC;
-    liveMode?: LiveModeProps["liveMode"];
+    liveMode: LiveModeProps["liveMode"];
     onLiveEvent?: LiveModeProps["onLiveEvent"];
+    options: IRefineContextOptions;
     children?: ReactNode;
 }