+++ packages/core/src/definitions/helpers/index.ts
@@ -6,3 +6,6 @@ export { queryKeys } from "./queryKeys";
 export { hasPermission } from "./hasPermission";
 export { routeGenerator } from "./routeGenerator";
 export { createTreeView } from "./treeView/createTreeView";
+export { humanizeString } from "./humanizeString";
+export { handleRefineOptions } from "./handleRefineOptions";
+export { redirectPage } from "./redirectPage";