+++ packages/react-table/test/index.tsx
@@ -25,7 +25,7 @@ export const TestWrapper: (
                     dataProvider={dataProvider ?? MockJSONServer}
                     routerProvider={MockRouterProvider}
                     resources={resources ?? [{ name: "posts" }]}
-                    disableTelemetry={true}
+                    options={{ disableTelemetry: true }}
                 >
                     {children}
                 </Refine>